requirements.txt
# AI 框架
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
langgraph>=0.0.20

# LLM 相关
openai>=1.0.0

# 数据处理
pandas>=2.0.0

# 数据存储
chromadb>=0.4.0
motor>=3.3.0  # MongoDB 异步驱动

# 数据采集
scrapy>=2.11.0
tushare>=1.3.0

# 前端界面
streamlit>=1.28.0
gradio>=4.0.0

# FastAPI 后端
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6  # 文件上传支持

# 文档解析
pdfplumber>=0.10.0  # PDF 解析（推荐）
PyPDF2>=3.0.0  # PDF 解析（备用）
python-docx>=1.0.0  # Word 文档解析

# 工具库
pydantic>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
typing-extensions>=4.8.0  # 类型注解支持